# Performance investigation and improvements for `random_agent`.