--- a/test/performance/get_random_agent.jl
+++ b/test/performance/get_random_agent.jl
@@ -7,7 +7,7 @@ mutable struct LabelledAgent <: AbstractAgent
 end
 
 n_agents = 100
-agents = [LabelledAgent(id, id<=n_agents/100) for id in 1:n_agents]
+agents = [LabelledAgent(id, id<=n_agents/3) for id in 1:n_agents]
 noremove_model = UnkillableABM(LabelledAgent)
 dict_model = ABM(LabelledAgent)
 fixed_model = FixedMassABM(agents)
@@ -32,19 +32,19 @@ end
 
 # All times are median
 # FixedMassABM
-@benchmark optimistic_random_agent(fixed_model, cond)
-@benchmark allocating_random_agent(fixed_model, cond)
-@benchmark random_agent(fixed_model, cond)
-@benchmark old_random_agent(fixed_model, cond)
+@benchmark optimistic_random_agent($fixed_model, $cond)
+@benchmark allocating_random_agent($fixed_model, $cond)
+@benchmark random_agent($fixed_model, $cond)
+@benchmark old_random_agent($fixed_model, $cond)
 
 # UnkillableABM
-@benchmark optimistic_random_agent(noremove_model, cond)
-@benchmark allocating_random_agent(noremove_model, cond)
-@benchmark random_agent(noremove_model, cond)
-@benchmark old_random_agent(noremove_model, cond)
+@benchmark optimistic_random_agent($noremove_model, $cond)
+@benchmark allocating_random_agent($noremove_model, $cond)
+@benchmark random_agent($noremove_model, $cond)
+@benchmark old_random_agent($noremove_model, $cond)
 
 # DictionaryABM
-@benchmark optimistic_random_agent(dict_model, cond)
-@benchmark allocating_random_agent(dict_model, cond)
-@benchmark random_agent(dict_model, cond)
-@benchmark old_random_agent(dict_model, cond)
+@benchmark optimistic_random_agent($dict_model, $cond)
+@benchmark allocating_random_agent($dict_model, $cond)
+@benchmark random_agent($dict_model, $cond)
+@benchmark old_random_agent($dict_model, $cond)