This is a draft PR intended for discussion and changes before pulling.

This PR addresses #742 by implementing two new functions for getting a random agent from an ABM as well as benchmarking them to find out which is most suitable. Additionally, the `random_agent` function is given a new kwarg `optimistic=false` controlling whether it should do a (normally) non-allocating "optimistic" search using `optimistic_random_agent` or the traditional method, renamed to `allocating_random_agent`.

`optimistic_random_agent` is implemented as follows: pick a random agent, does it satisfy the condition?. It does this until `n_attempts` is reached, at which point one of two things could potentially happen:
1. Fallback to (an implementation of) `allocating_random_agent` (as it currently does) and return a random agent.
2. Return nothing.

The former offers stronger correctness guarantees, the latter stronger performance guarantees (the function will never allocate, compilers like this, but it still is impure due to randomness so probably not such a big deal) while sacrificing correctness.

Additionally, this PR also implements what I claim is a more efficient `allocating_random_agent` (currently called `allocating_random_agent2`) which first applies an inplace filter to the collected agent IDs, then just picks one at random. Benchmarks can be found in `test/performance/get_random_agent.jl` and summarised below as:

```julia
# All times are median
# FixedMassABM
@benchmark optimistic_random_agent(fixed_model, cond) # 336 ns
@benchmark allocating_random_agent(fixed_model, cond) # 1450 ns
@benchmark allocating_random_agent2(fixed_model, cond) # 6100 ns

# UnkillableABM
@benchmark optimistic_random_agent(noremove_model, cond) # 158 ns 
@benchmark allocating_random_agent(noremove_model, cond) # 1283 ns
@benchmark allocating_random_agent2(noremove_model, cond) # 276 ns

# DictionaryABM
@benchmark optimistic_random_agent(dict_model, cond) # 454 ns
@benchmark allocating_random_agent(dict_model, cond) # 1592 ns
@benchmark allocating_random_agent2(dict_model, cond) # 882 ns
```

The results show that apart from for fixed-mass ABMs, the filter-first technique is an improvement. It allocates the same (number and size).