@jacobusmmsmit 2023-02-16
A key thing that I haven't mentioned is that it is very lenient condition that I am checking here. 50% of the agents have this condition satisfied. Speedup would be less for fewer agents.