@Tortar 2023-02-16
Cool idea to have a probabilistic search! I modified the old `allocating_random_agent`, now the ways are more or less of the same performance