@jacobusmmsmit 2023-02-16
If the performance difference between the two methods is consistently different between container types it might be worth dispatching on it.