@Tortar 2023-02-16
need to try one last optimization of `allocating_random_agent`, but I like the idea!