@Tortar 2023-02-17
On the other hand, with the optimistic search I don't know if the dispatching is worthy enough