@Tortar 2023-02-17
Also I'm finding a bit strange that the new way is so much better in one case, but worse in the others since the two functions both collects ids and so they then operate on the same data structures from that point on, while the `allocating_random_agent` is almost constant, so how can it be?