@Tortar 2023-02-17
wait a second, in the benchmark it shouldn't be

```
noremove_model = UnkillableABM(agents)
dict_model = ABM(agents)
fixed_model = FixedMassABM(agents)
```
?