@Tortar 2023-02-17
the new function should be much better than both! with the 10000 agents as above with 50% density I calculated

```julia
# FixedMassABM
6.265 μs
# UnkillableABM
6.066 μs 
# DictionaryABM
17.282 μs
```

Hope it's correct