@jacobusmmsmit 2023-02-17
That looks great, I had a go trying out `popat!` instead of swapping elements and using a decrementing `N` each time instead of calling `eachindex` each loop, but your method seems to be a few `ns` faster in each case.