@Tortar 2023-02-17
@jacobusmmsmit just to mention: `popat!` should become slower and slower than swapping and removing the last more and more the agent number gets greater