@Datseris 2023-02-17
Okay so what is the status quo here? Two new versions are offered, the old version was removed. Both versions are faster than the old?