@Tortar 2023-02-17
the last version has proved the fastest, that's why @jacobusmmsmit removed the others. The first version proposed by him was faster but not always, while this last one is (up to 20x) than the old one with all models.