@Datseris 2023-02-17
I admit I find it difficult to believe that a single `collect` and `shuffle!` is slower than `collect` and arbitrarily many calls to `rand` and `pop!`, but then again you tested it and the numnbers don't lie.

Please replace all calls to `model.rng` with `abmrng(mode)`, add a changelog entry and bump minor release up so that I can merge and tag this improvement !