@jacobusmmsmit 2023-02-17
I'm not 100% happy with this yet because `random_agent` now has 2 allocations. Not sure why. It's still faster (even if just a bit) than the original `random_agent` but the allocation is annoying.