@Datseris 2023-02-17
This is because you do not use `$` in the `@benchmark`. Instead of `@benchmark f(a, b)` you should write `@benchmark f($a, $b)`.