@jacobusmmsmit 2023-02-17
I think the key reason why it's faster is because `shuffle!` has to generate a permutation for the whole length of the vector every single time, whereas the new default implementation only generates `x` random numbers each time where `x <= length(vector)`.