@jacobusmmsmit 2023-02-17
Here are the benchmark results for a `Dict` ABM (the most common case) when `n_agents = 100` and $\mathbb{P}(match) = \frac{1}{3}$:
```julia
# Optimistic
BenchmarkTools.Trial: 10000 samples with 231 evaluations.
 Range (min … max):  398.268 ns … 735.571 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     483.048 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   484.860 ns ±  30.025 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                     ▂▂▂▄▅▆▇▆▅█▇▇▇▇▅▅▃▃▂▂▁                       
  ▁▁▁▁▁▂▂▂▂▂▃▃▄▄▅▆▇▇▇█████████████████████▇▇▆▆▆▄▄▄▃▃▂▂▂▂▂▂▁▁▂▁▁ ▄
  398 ns           Histogram: frequency by time          572 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# Allocating
BenchmarkTools.Trial: 10000 samples with 201 evaluations.
 Range (min … max):  412.517 ns …  22.619 μs  ┊ GC (min … max): 0.00% … 97.78%
 Time  (median):     476.990 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   503.189 ns ± 883.449 ns  ┊ GC (mean ± σ):  7.62% ±  4.24%

      ▁▆█▆▂                 ▃▃▂▁                                 
  ▁▁▂▄█████▆▃▂▂▂▁▁▁▁▁▁▁▁▂▃▆▇████▇▆▄▄▃▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  413 ns           Histogram: frequency by time          573 ns <

 Memory estimate: 896 bytes, allocs estimate: 1.

# New random_agent (with optimistic=false)
BenchmarkTools.Trial: 10000 samples with 199 evaluations.
 Range (min … max):  413.945 ns …  23.040 μs  ┊ GC (min … max): 0.00% … 97.67%
 Time  (median):     477.181 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   504.668 ns ± 902.871 ns  ┊ GC (mean ± σ):  7.76% ±  4.24%

      ▂██▅▃                ▂▃▄▂▁                                 
  ▁▁▂▅█████▆▃▂▂▂▁▁▁▁▁▁▁▁▃▄▇█████▆▅▄▄▃▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  414 ns           Histogram: frequency by time          576 ns <

 Memory estimate: 896 bytes, allocs estimate: 1.

# Old random_agent
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.383 μs … 439.625 μs  ┊ GC (min … max): 0.00% … 99.21%
 Time  (median):     1.558 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.610 μs ±   4.382 μs  ┊ GC (mean ± σ):  2.71% ±  0.99%

           ▅▂▃▃▆  ▃ ▁▅▂▂█▂▁ ▁                                  
  ▁▁▁▂▂▃▆▆███████▇███████████▆▄▅▂▂▃▂▁▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  1.38 μs         Histogram: frequency by time        1.95 μs <

 Memory estimate: 896 bytes, allocs estimate: 1.
```