@Tortar 2023-02-21
Sorry for the inexperience, but we should add it to the changelog under 5.7 and then change 5.7.0 in 5.7.1 in project.toml? Or put it under a new 5.8 and then change 5.7.0 in 5.8.0?