@Datseris 2023-02-25
Thanks guys! Alright, so, performance improvements increment _patch_ version. So this would go under 5.7.1. But now I'm merging as is, because there are other PRs going on, so there is no pressing need to change project.toml version in this PR. Bceause there are a lot of PRs, it is likely that we will release 5.8 anyways, so this should get a changelog entry under 5.8. 

(I'm not waiting for the changelog entry. merging away now, feel free to do followup PR for changelog)