@Datseris 2023-02-25
Oh what am I saying. This has to increment the **minor version**, not patch. Because it introduces a new API name, a new keyword for the function. So it anyways goes under 5.8...